Added right click and (hopefully) fixed issues)
Added right click exit. I think the problem with the compiler errors was that I didnt realize the moniter prefab for the laptop and the moniters was the same so each one needed its own thing. My fix was to create a new laptop moniter prefab from the original. Let me know if I broke everything.